requirements.txt
flask>=3.0
faster-whisper>=0.10
python-dotenv>=1.0
openai>=1.3
epitran>=1.2

